requirements.txt
websockets==10.3
python-dotenv==0.20.0
